refactor: co-locate BullMQ queue and worker with document feature module
Move queue definition and worker processor from centralized src/lib/queue/
to src/document/ to align with the feature-based architecture. This keeps
job infrastructure next to the business logic it serves and scales better
as more features add background jobs.
