+++ src/document/queue.ts
@@ -1,5 +1,5 @@
 import { Queue } from "bullmq";
-import { connection } from "../connection";
+import { connection } from "@/lib/queue/connection";
 
 export interface SendToTaxEntityJobData {
   orderId: string;
